Create compile time PlatformActivity base class to avoid dupicated code
